+++ relevanceai/cluster_report/cluster_report.py
@@ -7,6 +7,7 @@ Automated Cluster Reporting
 .. note::
     **Introduced in v1.0.0.**
 
+You can run cluster reporting as a standalone module.
 
 .. code-block::
 
@@ -23,7 +24,7 @@ Automated Cluster Reporting
     X = np.array(df['base_vector_'].tolist())
 
 
-    from relevanceai.cluster_report.cluster_report import ClusterReport
+    from relevanceai.cluster_report import ClusterReport
     from sklearn.cluster import KMeans
 
     N_CLUSTERS = 2
@@ -37,14 +38,66 @@ Automated Cluster Reporting
         model=kmeans
     )
 
-    internal_report = report.get_cluster_internal_report()
+    # JSON output
+    report.internal_report
+    
+    # Prettyprinted report of overall statistics
+    report.internal_overall_report
+
+
+You can also insert your own centroid vectors if you want them to be represented. 
+For example - you may want to measure off medoids (points in your dataset) instead of centroids
+(as opposed to points outside of your dataset).
+
+In the example below, we show how you calculate centroids or medoids for HDBSCAN
+
+.. code-block::
+
+    from relevanceai.cluster_report import ClusterReport
+
+    import hdbscan
+    clusterer = hdbscan.HDBSCAN()
+
+    cluster_labels = clusterer.fit_predict(X)
+    centroids = ClusterReport.calculate_centroids(X, cluster_labels)
+    # or if you want medoids
+    # medoids = ClusterReport.calculate_medoids(X, cluster_labels)
+
+    report = ClusterReport(
+        X=X, 
+        cluster_labels=cluster_labels, 
+        centroid_vectors=centroids
+    )
+    report.internal_overall_report
+
+.. code-block::
+
+    from relevanceai.cluster_report import ClusterReport
+
+    import hdbscan
+    clusterer = hdbscan.HDBSCAN()
+
+    cluster_labels = clusterer.fit_predict(X)
+    medoids = ClusterReport.calculate_medoids(X, cluster_labels)
+
+    report = ClusterReport(
+        X=X, 
+        cluster_labels=cluster_labels, 
+        centroid_vectors=centroids
+    )
+    report.internal_overall_report
 
 """
 
 import pandas as pd
 import numpy as np
+from relevanceai.integration_checks import is_hdbscan_available, is_sklearn_available
 from relevanceai.warnings import warn_function_is_work_in_progress
-from typing import Union, List, Dict, Any
+from typing import Union, List, Dict, Any, Optional
+import functools
+from warnings import warn
+from doc_utils import DocUtils
+from relevanceai.analytics_funcs import track
 
 try:
     from sklearn.metrics import (
@@ -55,13 +108,14 @@ try:
     from sklearn.metrics.pairwise import (
         pairwise_distances,
     )
+    from sklearn.cluster import MiniBatchKMeans, KMeans
     from sklearn.tree import _tree, DecisionTreeClassifier
-    from sklearn.cluster import KMeans
+    from sklearn.neighbors import NearestNeighbors
 except ModuleNotFoundError as e:
     pass
 
 
-class ClusterReport:
+class ClusterReport(DocUtils):
     """
     Receive an automated cluster reprot
 
@@ -83,17 +137,28 @@ class ClusterReport:
         The model to analyze. Currently only used
     num_clusters: Optional[int]
         The number of clusters. This is required if we can't actually tell how many clusters there are
-
+    outlier_label: Optional[str, int]
+        The label if it is an outlier
+    centroids: Union[list, np.ndarray]
+        The centroid vectors. If supplied, will use these. Otherwise, will try to infer them
+        from the model.
     """
 
+    @track
     def __init__(
         self,
         X: Union[list, np.ndarray],
         cluster_labels: List[Union[str, float]],
         model: KMeans = None,
         num_clusters: int = None,
+        outlier_label: Union[str, int] = -1,
+        centroids: Union[list, np.ndarray] = None,
+        verbose: bool = False,
     ):
         warn_function_is_work_in_progress()
+
+        self._typecheck_model(model)
+
         if isinstance(X, list):
             self.X = np.array(X)
         else:
@@ -103,9 +168,31 @@ class ClusterReport:
         else:
             self.cluster_labels = cluster_labels
         self.num_clusters = (
-            len(cluster_labels) if num_clusters is None else num_clusters
+            len(set(cluster_labels)) if num_clusters is None else num_clusters
         )
         self.model = model
+        self.outlier_label = outlier_label
+        self._typecheck_centroid_vectors(centroids)
+        self._centroids = centroids
+        self.verbose = verbose
+
+    def _typecheck_centroid_vectors(
+        self, centroid_vectors: Optional[Union[list, Dict, np.ndarray]] = None
+    ):
+        if isinstance(centroid_vectors, (list, np.ndarray)):
+            warn(
+                "Centroid vectors are a list. Assuming they are in the order of the cluster labels."
+                + "To specify which vectors mapped to which label, place in the format of "
+                + "{cluster_label: centroid_vector}."
+            )
+
+    def _typecheck_model(self, model):
+        if is_hdbscan_available():
+            return
+        if is_sklearn_available():
+            if isinstance(model, (KMeans, MiniBatchKMeans)):
+                return
+        warn("Model not directly supported. Will try to infer.")
 
     @staticmethod
     def summary_statistics(array: np.ndarray, axis=0):
@@ -178,28 +265,37 @@ class ClusterReport:
     def get_z_score(value, mean, std):
         return (value - mean) / std
 
-    # TODO: Implement caching
-    def get_centers(self):
-        # Here we add support for both RelevanceAI cluster models
-        # but also regular sklearn cluster models
+    @functools.lru_cache(maxsize=128)
+    def get_centers(self, output_format="array"):
         if hasattr(self.model, "cluster_centers_"):
             return self.model.cluster_centers_
         elif hasattr(self.model, "get_centers"):
             return self.model.get_centers()
+        elif self._centroids is not None:
+            if output_format == "array":
+                if isinstance(self._centroids, dict):
+                    return np.array(list(self._centroids.values()))
+                else:
+                    return self._centroids
+            else:
+                return self._centroids
         else:
-            import warnings
-
-            warnings.warn("Missing centroids.")
+            if self.verbose:
+                print(
+                    "No centroids detected. We recommend including centroids to get all stats."
+                )
             return
 
-    def get_cluster_internal_report(self):
+    @property  # type: ignore
+    @functools.lru_cache(maxsize=128)
+    def internal_report(self):
         """
         Provide the standard clustering report.
         """
         self.X_silhouette_scores = silhouette_samples(
             self.X, self.cluster_labels, metric="euclidean"
         )
-        self.cluster_internal_report = {
+        _internal_report = {
             "overall": {
                 "summary": ClusterReport.summary_statistics(self.X),
                 "davies_bouldin_score": davies_bouldin_score(
@@ -218,23 +314,26 @@ class ClusterReport:
                 "silhouette_score": {},
             },
         }
-        self._store_basic_centroid_stats(self.cluster_internal_report["overall"])
+        self._store_basic_centroid_stats(_internal_report["overall"])
 
-        label_value_counts = np.unique(self.cluster_labels, return_counts=True)
+        labels, counts = np.unique(self.cluster_labels, return_counts=True)
+        if self.verbose:
+            print("Detected the cluster labels:")
+            print(labels)
 
         cluster_report = {"frequency": {"total": 0, "each": {}}}
 
-        for i, cluster_label in enumerate(label_value_counts[0]):
+        for i, cluster_label in enumerate(labels):
             cluster_bool = self.cluster_labels == cluster_label
 
             specific_cluster_data = self.X[cluster_bool]
             other_cluster_data = self.X[~cluster_bool]
 
-            cluster_frequency = label_value_counts[1][i]
+            cluster_frequency = counts[i]
             cluster_report["frequency"]["total"] += cluster_frequency
             cluster_report["frequency"]["each"][cluster_label] = cluster_frequency
 
-            self.cluster_internal_report["each"]["summary"][
+            _internal_report["each"]["summary"][
                 cluster_label
             ] = ClusterReport.summary_statistics(self.X)
 
@@ -245,29 +344,33 @@ class ClusterReport:
                 specific_cluster_data
             )
 
-            squared_errors = np.square(
-                np.subtract(
-                    [self.get_centers()[i]] * len(self.X[cluster_bool]),
-                    self.X[cluster_bool],
-                )
-            )
             if self.has_centers():
 
                 grand_centroid = self.X[cluster_bool].mean(axis=0)
-                self.cluster_internal_report["overall"]["grand_centroids"].append(
-                    grand_centroid
+
+                centroid_vector = self._get_centroid_vector(
+                    i, cluster_label, grand_centroid
                 )
+
+                squared_errors = np.square(
+                    np.subtract(
+                        [centroid_vector] * len(specific_cluster_data),
+                        specific_cluster_data,
+                    )
+                )
+
+                _internal_report["overall"]["grand_centroids"].append(grand_centroid)
 
                 center_stats[
                     "distance_from_centroid"
                 ] = self.get_distance_from_centroid(
-                    specific_cluster_data, self.get_centers()[i]
+                    specific_cluster_data, centroid_vector
                 )
 
                 center_stats[
                     "distance_from_centroid_to_point_in_another_cluster"
                 ] = self.get_distance_from_centroid_to_another(
-                    other_cluster_data, self.get_centers()[i]
+                    other_cluster_data, centroid_vector
                 )
 
                 center_stats["distances_from_grand_centroid"] = pairwise_distances(
@@ -289,27 +392,23 @@ class ClusterReport:
                 center_stats["by_features"][
                     "overall_z_score"
                 ] = ClusterReport.get_z_score(
-                    self.get_centers()[i],
-                    self.cluster_internal_report["overall"]["summary"]["mean"],
-                    self.cluster_internal_report["overall"]["summary"]["std"],
+                    centroid_vector,
+                    _internal_report["overall"]["summary"]["mean"],
+                    _internal_report["overall"]["summary"]["std"],
                 )
 
                 center_stats["by_features"]["z_score"] = ClusterReport.get_z_score(
-                    self.get_centers()[i],
-                    self.cluster_internal_report["each"]["summary"][cluster_label][
-                        "mean"
-                    ],
-                    self.cluster_internal_report["each"]["summary"][cluster_label][
-                        "std"
-                    ],
+                    centroid_vector,
+                    _internal_report["each"]["summary"][cluster_label]["mean"],
+                    _internal_report["each"]["summary"][cluster_label]["std"],
                 )
 
                 center_stats["by_features"][
                     "overall_z_score_grand_centroid"
                 ] = ClusterReport.get_z_score(
                     grand_centroid,
-                    self.cluster_internal_report["overall"]["summary"]["mean"],
-                    self.cluster_internal_report["overall"]["summary"]["std"],
+                    _internal_report["overall"]["summary"]["mean"],
+                    _internal_report["overall"]["summary"]["std"],
                 )
 
                 center_stats[
@@ -320,12 +419,8 @@ class ClusterReport:
 
                 center_stats["by_features"]["z_score_grand_centroid"] = (
                     grand_centroid
-                    - self.cluster_internal_report["each"]["summary"][cluster_label][
-                        "mean"
-                    ]
-                ) / self.cluster_internal_report["each"]["summary"][cluster_label][
-                    "std"
-                ]
+                    - _internal_report["each"]["summary"][cluster_label]["mean"]
+                ) / _internal_report["each"]["summary"][cluster_label]["std"]
 
                 # this might not be needed
                 center_stats["overall_z_score"] = ClusterReport.summary_statistics(
@@ -342,39 +437,69 @@ class ClusterReport:
                     center_stats["by_features"]["z_score_grand_centroid"]
                 )
 
+                # squared errors are calculted by the centroids
                 center_stats["squared_errors"] = ClusterReport.summary_statistics(
                     squared_errors, axis=2
                 )
 
-            squared_errors_by_col = {}
+                squared_errors_by_col = {}
 
-            for f in range(len(squared_errors[0])):
-                squared_errors_by_col[f] = ClusterReport.summary_statistics(
-                    squared_errors[:, f], axis=2
-                )
+                for f in range(len(squared_errors[0])):
+                    squared_errors_by_col[f] = ClusterReport.summary_statistics(
+                        squared_errors[:, f], axis=2
+                    )
 
-            center_stats["by_features"]["squared_errors"] = squared_errors_by_col
-            self.cluster_internal_report["each"]["centers"][
-                cluster_label
-            ] = center_stats
-            self.cluster_internal_report["each"]["silhouette_score"][
+                center_stats["by_features"]["squared_errors"] = squared_errors_by_col
+
+                _internal_report["each"]["centers"][cluster_label] = center_stats
+
+            _internal_report["each"]["silhouette_score"][
                 cluster_label
             ] = ClusterReport.summary_statistics(
                 self.X_silhouette_scores[cluster_bool], axis=2
             )
 
         if self.has_centers():
-            self.cluster_internal_report["overall"]["dunn_index"] = (
-                min(
-                    c["distance_from_centroid"]["min"]
-                    for c in self.cluster_internal_report["each"]["centers"].values()
-                )
-                / self.cluster_internal_report["overall"][
-                    "centroids_distance_matrix"
-                ].max()
+
+            min_centroid_distance = min(
+                c["distance_from_centroid"]["min"]
+                for c in _internal_report["each"]["centers"].values()
             )
 
-        return self.cluster_internal_report
+            max_centroid_distance = _internal_report["overall"][
+                "centroids_distance_matrix"
+            ].max()
+            _internal_report["overall"]["dunn_index"] = self.dunn_index(
+                min_centroid_distance, max_centroid_distance
+            )
+
+        return _internal_report
+
+    def _get_centroid_vector(
+        self, i: int = None, cluster_label: int = None, default_vector=None
+    ):
+        # If this is an outlier, we will automatically default to the grand centroid
+        if cluster_label == self.outlier_label:
+            if self.verbose:
+                print(
+                    "Outlier labels detected. Using the grand centroid for the outlier label."
+                )
+            return default_vector
+        centers = self.get_centers()
+        if isinstance(centers, (list, np.ndarray)):
+            centroid_vector = centers[i]  # type: ignore
+        elif isinstance(centers, dict):
+            try:
+                centroid_vector = centers[cluster_label]
+            except KeyError:
+                warn("cluster label not detected in centroid vectors")
+                centroid_vector = default_vector
+        else:
+            raise ValueError("Centroid vector needs to be a list or a dictionary.")
+        return centroid_vector
+
+    def dunn_index(self, min_distance_from_centroid, max_centroid_distance):
+        return min_distance_from_centroid / max_centroid_distance
 
     def has_centers(self):
         return self.get_centers() is not None
@@ -382,14 +507,84 @@ class ClusterReport:
     def _store_basic_centroid_stats(self, overall_report):
         """Store"""
         if self.has_centers():
-            overall_report["centroids"] = self.get_centers()
+            centroids = self.get_centers()
+            overall_report["centroids"] = centroids
             overall_report["centroids_distance_matrix"] = pairwise_distances(
-                self.get_centers(), metric="euclidean"
+                centroids, metric="euclidean"
             )
             overall_report["grand_centroids"] = []
             overall_report["average_distance_between_centroids"] = (
                 overall_report["centroids_distance_matrix"].sum(axis=1) - 1
             ) / self.num_clusters
+
+    @property
+    def davies_bouldin_score(self):
+        return pd.DataFrame(
+            self.subset_documents(
+                ["davies_bouldin_score"], [self.internal_report["overall"]]
+            )
+        )
+
+    @property
+    def calinski_harabasz_score(self):
+        return pd.DataFrame(
+            self.subset_documents(
+                ["calinski_harabasz_score"], [self.internal_report["overall"]]
+            )
+        )
+
+    @property
+    def internal_overall_report(self):
+        """
+        View the internal overall report.
+        """
+        # TODO: Figure out a better way to present this than having index in the middle
+        if not self.has_centers():
+            metrics = pd.DataFrame(
+                self.subset_documents(
+                    ["davies_bouldin_score", "calinski_harabasz_score"],
+                    [self.internal_report["overall"]],
+                )
+            )
+            overall_df = pd.DataFrame(self.internal_report["overall"])[
+                ["summary", "silhouette_score"]
+            ]
+            return pd.concat([metrics, overall_df.reset_index()], axis=1).fillna(" ")
+        else:
+            metrics = pd.DataFrame(
+                self.subset_documents(
+                    ["davies_bouldin_score", "calinski_harabasz_score", "dunn_index"],
+                    [self.internal_report["overall"]],
+                )
+            )
+
+            overall_df = pd.DataFrame(
+                self.get_fields(
+                    ["overall.summary", "overall.silhouette_score"],
+                    self.internal_report,
+                )
+            ).T
+            overall_df.columns = ["summary", "silhouette_score"]
+            return pd.concat([metrics, overall_df.reset_index()], axis=1).fillna(" ")
+
+    @staticmethod
+    def calculate_centroids(X, cluster_labels):
+        """Calculate the centes"""
+        centroid_vectors = {}
+        for label in cluster_labels:
+            centroid_vectors[label] = X[cluster_labels == label].mean(axis=0)
+        return centroid_vectors
+
+    @staticmethod
+    def calculate_medoids(X, cluster_labels):
+        centroids = ClusterReport.calculate_centroids(X, cluster_labels)
+        medoids = {}
+        nbrs = NearestNeighbors(n_neighbors=1, algorithm="ball_tree").fit(X)
+        medoid_indexes = nbrs.kneighbors(
+            np.array(list(centroids.values())), n_neighbors=1, return_distance=False
+        )
+        medoids = X[medoid_indexes]
+        return medoids
 
     def get_class_rules(self, tree: DecisionTreeClassifier, feature_names: list):
         self.inner_tree: _tree.Tree = tree.tree_